# game-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Strategy game backend with users, villages, resources, buildings, and units.",
        "title": "Game API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/units": {
            "get": {
                "description": "Zwraca listę jednostek i ich liczebność w wybranej wiosce",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "units"
                ],
                "summary": "Pobierz jednostki w wiosce",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID wioski",
                        "name": "village_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Missing or invalid village_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/units/recruit": {
            "post": {
                "description": "Rekrutuje jednostki w wiosce, o ile użytkownik posiada wystarczające zasoby.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "units"
                ],
                "summary": "Rekrutuj jednostki",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID wioski",
                        "name": "village_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Typ jednostki (spearman, swordsman, archer)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Liczba rekrutowanych jednostek",
                        "name": "count",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Missing params or invalid unit type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Not enough resources or village not yours",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/villages": {
            "get": {
                "description": "Zwraca listę wiosek należących do zalogowanego użytkownika",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "villages"
                ],
                "summary": "Pobierz wszystkie wioski użytkownika",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.Village"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Tworzy nową wioskę, jeśli użytkownik spełnia wymagania Townhallu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "villages"
                ],
                "summary": "Utwórz nową wioskę",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nazwa nowej wioski",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Missing village name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "You need higher Townhall level",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/villages/{id}": {
            "put": {
                "description": "Aktualizuje nazwę wioski, jeśli należy do użytkownika",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "villages"
                ],
                "summary": "Zmień nazwę wioski",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Village ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nowa nazwa wioski",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid village ID or missing name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Village not yours",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Usuwa wioskę użytkownika, wraz z powiązanymi danymi",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "villages"
                ],
                "summary": "Usuń wioskę",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Village ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid village ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Village not yours",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.Village": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}